+++ mne/viz/tests/test_epochs.py
@@ -225,7 +225,9 @@ def test_plot_psd_epochs():
     """Test plotting epochs psd (+topomap)."""
     epochs = _get_epochs()
     epochs.load_data()
-    epochs.plot_psd()
+    epochs.plot_psd(average=True, spatial_colors=False)
+    epochs.plot_psd(average=False, spatial_colors=True)
+    epochs.plot_psd(average=False, spatial_colors=False)
     pytest.raises(RuntimeError, epochs.plot_psd_topomap,
                   bands=[(0, 0.01, 'foo')])  # no freqs in range
     epochs.plot_psd_topomap()